Mobile badge should show capped, language-converted number
The badge starts out with the correctly capped number, but
when an item is marked as read/unread it updates itself again
according to the counter. The counter, however, has raw count
(on purpose) which is then displayed uncapped.

This commit fixes that and also makes sure we use the
mw.language.convertNumber() method to display the number
correctly per the interface language.

Depends on Echo change I4d7f406b05a195972dca0d2088

Change-Id: I9d934c7b91e6bbc0b4140c8b3a707e6ebff9000a
Depends-On: I4d7f406b05a195972dca0d2088bde2ff739d313d