Toast should be centered on mobile devices
Note this also applies border-box to the desktop version
of Minerva when styled via mw-notification (which has the
same problem) and to drawers.

Drawers are already styled with border-box by virtue of
inheriting and not overriding the isBorderBox property of
the view class (no existing Drawers reset this at time of
writing)

The `drawer` class should not be in this file and this will
be cleaned up in a follow up commit.

Bug: T144828
Change-Id: I050f91a41bdddcaad261e6b47dfe0ef95c7a566b