Update patch set 20
Patch Set 20: Code-Review+1

Jdlrobson: Annoying timing-related failure aside, I've tested this change locally and it LGTM. I'm using a MWV instance with the same setup (and the same Math configuration) as our staging server.

I'm happy with this change, though I'd note that the image dimensions reported by the equation renderer(s) are off, causing the page to reflow when an equation is lazily-loaded. IMO this should be fixed in the extension itself.

Patch-set: 20
Label: Code-Review=+1