Update patch set 6
Patch Set 6:

I agree. I meant for this specific patch. And we have a policy of tracking all work done in the sprint, in the sprint. That's why I mentioned putting it in the sprint. I think this is especially warranted as it's a fix for our breaking browser tests.

Patch-set: 6