Avoid calling mw.msg with variables
In order to make messages grepable and the sources parseable to extract
which sources use which messages, mw.msg needs to be called always with
string literals.

This commit changes uses of mw.msg( variable ) to use a string literal
instead. In most cases the refactoring is passing around the translated
string instead of the message key.

This fixes will be enforced by the resource-modules linter once added to
the extension.

Bug: T146748
Change-Id: I6bb8204b41d5ff43b0f0deb4f39daeebe2d23912