Update patch set 1
Patch Set 1: Code-Review-1

The inconsistency in whitespace between the heading or tagline and the page actions bar is caused by the tagline having a 10px bottom margin. If you remove this bottom margin, then you can increase the bottom padding of the container to 3.6 em regardless of its content.

I'd recommend the above rather than this solution as it removes the inconsistency in styles directly rather than indirectly.

Patch-set: 1
Label: Code-Review=-1