Import mw.ForeignApi to local scope instead of directly using it
This allows the linter to track where ForeignApi is in the dependencies
and complain if the config is not set so that ForeignApi loads before
JSONPForeignApi

Culprit of the not-trackability was
  return mw.ForeignApi.prototype.ajax.call( this, data, options );

That line is imposible to analyze statically and determine which of [mw,
ForeignApi, prototype, ajax, call] are actually defined global
namespaces/variables.

This change changes no semantics on the code.

Bug: T146748
Change-Id: Iae6d52a33785f83adaf72f8101feed6796b3b8ba