Update patch set 8
Patch Set 8:

> check experimental

I'm getting to a halt where there are some things to add to the linter like understanding the new require/exports pattern that's incompatible with node's one.

And I have to ask about if we can include the resource module definitions to extension.json in the case of visual editor or not.

Patch-set: 8