Update patch set 1
Patch Set 1:

> mobile-frontend-overlay-close only needs to be in mobile.overlays
 > since all overlays should inherit from that module. Duplicating it
 > like this will unnecessarily bloat the JS payload.

I think we should be striving for clarity and linting (source uses a message, then config of that source has that message on messages) and leave optimization for ResourceLoader to de-duplicate the messages if needed to.

Making this kind of manual optimizations on a 1700LOC JSON configuration is what makes it very hard to understand or maintain.

I've made a few quick tests with the previous patch, and this one, and the initial payloads are not "bloated", they seem the same to me.

I've tried initial load, open editor, and open image overlay.

The network requests for the three cases showed the same numbers on the network inspector on Chrome.

Example: pre patch: https://i.imgur.com/w0dzn2r.png and with this patch: https://i.imgur.com/w0dzn2r.png

So I'm not even sure that argument really applies. I'd love some concrete evidence to be able to work on it and move forward.

Patch-set: 1
Label: Verified=0