Update patch set 4
Patch Set 4:

TIL that the Resource Loader (the "RL") guarantees that modules aren't loaded more than once but it doesn't guarantee that sub-resources, like messages and templates, aren't.

In the case of messages: an RL module can define a list of message keys and that list can contain duplicates. Eventually, the list is converted to a "message blob" – a dictionary whose keys are the message keys and whose values are the "plain" (see `Message#plain`) messages in `MessageBlobStore#generateMessageBlob`. Because of this conversion, an RL module's messages are guaranteed to be unique.

However, message blobs are fetched per module and delivered as part of that module's definition (see `ResourceLoader::makeLoaderImplementScript`). So, while an RL module's messages are guaranteed to be unique, an RL response can contain one or modules all with the same message blobs.

Patch-set: 4