Move mobile-frontend-editor-anon to where it is used
This message was declared in skins.minerva.editor module where it is not
used, and mobile.editor.overlay where it is used doesn't have this
message (nor any of it's direct dependencies).

This message ended up being available by mere coincidence. This patch
moves it to where it is used (only place in the codebase)

Bug: T146748
Change-Id: I8af14098e00f78cfd4a25cc893376d1e25858276