Extract extra device detection from MobileContext
... and break it into discrete units.

Additional device detection logic exists in MobileContext#getAMF and
whether the request is sent from a mobile or tablet device: in the case
of #getAMF, environment variables set by the Apache Mobile Filter
module; in the case of #getDevice, a configurable request header.

Patrick Reilly's IDeviceDetector interface is a good start for a
boundary for all device detection code. However, in order to support the
cases above, it needs to be extended to accept more application state.

Changes:
* Create MobileFrontend\Devices\DeviceDetector from IDeviceDetector.
* Create DeviceProperties, a simple DTO, from IDeviceProperties.
* Add MobileFrontend\Devices\AMFDeviceDetector and
  CustomHeaderDeviceDetector, which cover the above cases.
* Create UADeviceDetector from IDeviceProperties, which covers the
  existing case

Bug: T143891
Change-Id: I8f5df2b8af407c2ec6763f2c2d7d99d3342065c1