Hygiene: Use `arr.push(var)` instead `arr = arr.concat([var])`
It's the semantically correct method to use, easier to read, runs faster,
uses less memory, and requires less keystrokes!

As in most languages, Array#push() adds items to an existing array.

Array#concat, on the other hand, is a utility to merge two or more arrays
into a new array. It creates a copy of the first array and adds items to
that copy. Aside from the extra copy it's also less intuitive in usage
because as a developer you first have to put the new items in another
temporary array, and, afterwards, overwrite the original variable.

Two extra object creations, and two extra objects dereferenced.

Change-Id: I7979f2cb10a1ddb7c8a62fb3fd0299748da3969a