Update patch set 2
Patch Set 2:

I can't reproduce the problem you're seeing. My local wiki only has the title "devwiki" with "Privacy" and "Desktop" links, no CC attribution. Can you describe the differences between stable and beta that you're seeing (i.e. which style rules are (not) being applied)? And are you certain your html/wikitext is correct? It looks like on production the license is not part of an hlist-separated like yours is in the screenshot.

Patch-set: 2