Add DeviceDetectorService class
In order to remove knowledge about device detection from MobileContext,
including the order of the methods we use to detect devices, it needs to
be moved somewhere: MobileFrontend\Devices\DeviceDetectorService.

If device detection is enabled, then DeviceDetectorService iterates
through a set of child device detectors, the default set of which are
defined in DeviceDetectorService::factory.

Changes:
* Extract DeviceDetectorService from MobileContext#isMobileDevice and
  #getDevice.
* Make MobileContext#isMobileDevice use DeviceDetectorService
* Remove MobileContext#getAMF and #getDevice.

Bug: T143891
Change-Id: I09547448f7eb1a56cea718a6751de227d2890987