Update patch set 1
Patch Set 1:

I'm not sure I understand the question.

In the commit message I mention how I could have removed the HTML for the edit link but the MobileFormatter has no knowledge of the editable state of a page.

I think the CTA should actually be rendered at the skin level but that's a much larger commit and would be one outcome of https://phabricator.wikimedia.org/T145219

Patch-set: 1