Update patch set 4
Patch Set 4:

You may be right but I didn't write this code and there are no comments so all I can go by is the visual difference with and without it.

If you are 100% sure this is no longer necessary I'll happily remove it but I'd be worried if we're unsure.

Patch-set: 4
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0