Update patch set 1
Patch Set 1:

It seems this is a behavioural regression rather than a unbreak now type regression (to be clear it seems anons cannot edit articles which is would be much more alarming! The protected page [1] is still not editable.

Previously we showed a toast when you clicked the edit icon however now the edit icon is not visible.

I'm not quite sure why this patch made this happen, but is it possible we were showing an edit icon before and not loading the JavaScript - this seems like a worse situation to revert to if true, so I'd rather we investigated this further and identified why this broke it before merging in case we are reverting to an even worse state. 

[1] https://en.m.wikipedia.beta.wmflabs.org/wiki/Selenium_protected_test_2

Patch-set: 1