Update patch set 1
Patch Set 1:

So it looks like the following happened:

Previously:
* Edit icon was shown if and only if the page was editable via JS.
* isCurrentPageEditable would have been better named CurrentPageEditableByCurrentUser as it was only checking if a user could edit the page.
* We loaded JavaScript code if the edit icon was present, even if that was just to show a toast saying the page wasn't editable.

With that in mind I'm not against reverting this, but I should note, the toast in the previous state is a little misleading 

If $wgGroupPermissions['*']['edit'] = false; - ie we disable editing by anon users

when the user clicks the edit icon they see a toast saying "This page is protected to prevent vandalism". This doesn't seem helpful, as a user can click edit on any page and see this toast when probably they should be advised to login. However if we don't revert this patch they don't see the edit icon at all.

Can you update your commit message with the "why"?

I'll have a go at fixing this this morning and will let Baha make a decision on whether it makes sense to revert it or not by the end of the day.

The downside of reverting is we break editing on the user page again which seems more important to me then showing a misleading toast to someone who is not allowed to edit.

Patch-set: 1