Update patch set 1
Patch Set 1:

> We're close enough to a branch cut that I'm inclined to Revert Now Fix Later™.
This is exactly why I'm suggesting not reverting. If we revert we introduce a feature regression for a behaviour regression into the cluster. We break editing of user pages for the sake of showing a edit icon which cannot be used.

The latter seems the lesser evil so I feel like this is one of those rare situations where reverting is actually the wrong thing to do.

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0