Blank user pages should be editable (attempt #2)
Introduce two methods
* isCurrentPageEditableByUser and isCurrentPageContentModelEditable
to separate the logic that existed in isCurrentPageEditable.

To preserve existing behaviour of editing protected pages,
isCurrentPageEditableByUser is used to determine whether to load the
skins.minerva.newusers module. This code is not loaded when a new user
signs up to edit a  protected page to avoid confusing
users by telling them to click a button that doesn't work.

A hidden and edit-link class is added to editor links on the user page.
skins.minerva.editor will operate on these links as well as section
edit pencil icon and page action edit icon. When initialised the hidden
class is removed and revealed to the user

To test:
* Protected pages should show edit button and when clicked will show a
toast
* User pages are now editable

Note:
* It does not support $wgMFAllowNonJavaScriptEditing

Bug: T143854
Change-Id: I47a4ad8ce0dd16dbf09a7a769ac0d0e4ec2a7c10