Make mobile.drawers depend on mobile-frontend-overlay-close
Explicitly, previously it was an implicit dependency and the message was
in place just by chance.

This patch makes mobile.drawers depend on mobile.overlays that defines
mobile-frontend-overlay-close.

Another (dismissed) option could be adding mobile-frontend-overlay-close
directly to mobile.drawers messages, but that would result in duplicate
messages in the output bundles.

Bug: T146748
Change-Id: I9f9f3cd85a132609c9720a0ecc2c1ddba4d64994