Collapse sections by default
A global function mfTempOpenSection is added to the top of the page
and all heading elements have click handlers bound to it.

As soon as skins.minerva.toggling is enabled these will be unbound
and the global destroyed and it will be business as usual.

In the mean time for eager readers it will be possible to expand sections.
It will not be possible to collapse them again (easily added but not worth
the additional complexity)

The experience remains the same as before for non-js users thanks to
client-nojs and toggle.js will honour any existing sections that are
open

Bug: T126825
Change-Id: Ia342b3ddfbf0a419122aa0de282310396fa264e7