Update patch set 5
Patch Set 5:

> I've set `$wgGroupPermissions['*']['edit'] = false;` in the config file then went to a user page that hasn't been created yet (the user account has been created). When I clicked on the edit link, the editor didn't show up.

This is because https://gerrit.wikimedia.org/r/#/c/313470/ didnt get squashed into this for some reason

> I also think that we should support editing when `$wgMFAllowNonJavaScriptEditing = true;`.

As I mentioned in the earlier patchset, the code for the user cta is currently bundled up in the MobileFormatter code which makes it tricky to do anything like this and I don't think its wise to couple MobileFormatter to the experimental feature MFAllowNonJavaScriptEditing . If it was in the skin it would be a lot easier, but if we want to do that, I would estimate it's about an 8 point story to move that into the skin and we should redefine the task.

Given the lack of activity on MFAllowNonJavaScriptEditing it's also tempting to remove that code altogether in a follow up.

I've updated the commit message to mention this

Patch-set: 5