Update patch set 5
Patch Set 5:

> As fro the previous case, this time I didn't see the edit link after I've set `$wgGroupPermissions['*']['edit'] = false;` in the config file then went (as anon) to a user page that hasn't been created yet (the user account has been created).

I'm not sure I follow if this is a -1 comment or a request for clarification.

I followed the logic of the section edit icons. Currently in this situation, for main namespace pages, we show the edit at the top of the page in page actions,  but we hide edit pencils next to section headings. We show a toast when you click the page action saying they cannot edit (which seems like a bad user experience that we should probably revisit - surely we should show them an invitation to sign up/login if someone has expressed a desire to login).

I followed the section heading logic, as it didn't seem helpful to show something that the user cannot interact with, and probably shouldn't (given the message "This page should be created and edited by <username>")

It's easy to switch the logic to show the button and make it show a toast on click, if you feel strongly that this is the right behaviour and want to run that by Nirzar. Let me know either way.

Patch-set: 5
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0