Get rid of footer margin collapsing on Special:MobileDiff
The bug affects Safari on iOS 9.3.

Add transparent boder bottom to the enclosing element so that the
footer margin bottom is not collapsed. An alternative would be
to change the margin bottom to padding bottom, but that would
make padding bottom calculation more complex because we need to
take the existing padding into account as well.

Bug: T144579
Change-Id: I3e3fdeee9c9f02cccb61914ae4c2940370cc193f