Update patch set 5
Patch Set 5:

First of all, thanks for your review. You're not overthinking anything.

>  Something feels off about this class [RequestHeaderDeviceDetector] having `isMobileDevice` and `isTabletDevice` methods when this patch also creates the `DeviceProperties` class for this purpose. Is it possible [DeviceDetector child classes] should extend that class? Maybe the `DeviceProperties` class is unnecessary and should be replaced by a simple dictionary (it is, after all, just a 2 entry dictionary)? Perhaps I'm just overthinking it though.

As I noted in `DeviceProperties` DocBlock, it's a Data Transfer Object ("DTO"). I consider DTOs superior to dictionaries for passing data around because:

1. They have a type.
2. You can't access a key that doesn't exist – accidentally or intentonally – as your program will grind to a halt, i.e. currently, PHP will trigger a fatal error when a method doesn't exist.

The above said, I prefer to create functions whose results are single values of a built-in type.

You'll note that in [T143891](https://phabricator.wikimedia.org/T143891) I originally intended the device detector service to answer two separate questions. However, since we'd always ask "is this request sent by a UA on a mobile device" and "… on a tablet device" together, I felt that Patrick Reilly's original design made sense.

> Would it perhaps make even more sense for `DeviceDetector` to be an abstract base class rather than an interface, in which you could merge `DeviceProperties`? The constructor could then call `detectDeviceProperties` (or assume its responsibilities), giving access to `isMobileDevice` and `isTabletDevice` upon instantiation.

The API would be the same but I'm not sure that it would make more sense.

`RequestHeaderDeviceDetector#isMobileDevice` and `#isTabletDevice` were misnomers in that they didn't reflect what they did – detect, or try to detect, properties of the device that the UA is sending requests from from the User-Agent request header. I've corrected this mistake.

More generally, if I'm creating an API, then I define an interface, not only because of the last letter of the acronym but because interfaces place only one restriction on their implementation: all members of the interface must be implemented by the class.

I'll add to this below.

> Again, it's highly likely I'm just overthinking it, but maybe there was a bit of overengineering done here?

I've thought about that word a lot since I read it. I believe that in this change I haven't over-engineered anything but that in the next change I have. I'm going to add an integration test and replace `CompositeDeviceDetector` as we're highly unlikely to require the flexibility that it introduces.

> More questions:
>
> - Are there scenarios in which one would call `detectDeviceProperties` multiple times against the same `DeviceDetector` object? If not, this sounds like it should be a static function OR done in a constructor.

Currently, no. I can conceive of such scenarios but that these scenarios are allowed for is a consequence of the design and not a goal.

Again, consider the question that the device detector answers:

> "What are the properties of the device that the UA sent this request from?"

I think that `DeviceDetector#detectDeviceProperties` captures this question well, i.e. that the answer to the question is a function of the request.

By the way, I've tried to make it clear that second parameter of `detectDeviceProperties` is necessary only because `WebRequest` is incomplete.

I use constructors are primarily to inject dependencies. The `AMFDeviceDetector` doesn't have any dependencies and, consequently, should be a function – not even a static function! However, the `RequestHeaderDeviceDetector` does depend on configuration.

> - Is there a specific reason for splitting the `AMFDeviceDetector` and `RequestHeaderDeviceDetector` classes? Do they require different constructor

The Single Responsibility Principle. Indeed, I've split  `RequestHeaderDeviceDetector` into `UADeviceDetector` and `CustomHeaderDeviceDetector` since your review.

Patch-set: 5