Update patch set 6
Patch Set 6:

>> - Is there a specific reason for splitting the `AMFDeviceDetector` and `RequestHeaderDeviceDetector` classes? Do they require different constructor
> The Single Responsibility Principle. Indeed, I've split  `RequestHeaderDeviceDetector` into `UADeviceDetector` and `CustomHeaderDeviceDetector` since your review.
My only thought here is that when a client (i.e. one who needs to detect a device, not an end-user) needs to check for a device, they may not care or know whether that device is detected via UA, AMF, or custom headers. A more generic Q&A of "what's the device?" "here you go" seems like a more applicable use case, which would necessitate a more generic class. That being said, I have yet to review the other patches in the associated task, so if that is taken care of there, feel free to simply respond with such.

Patch-set: 6
Reviewer: Jhobs <1948@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0