Update patch set 1
Patch Set 1:

FWIW I find section collapsing really useful. For me it serves a different purpose to table of contents. I use it to keep track of sections I want to read and have read and it allows me to organise the page so that it's not cluttered with information I'm not interested in (I use it on tablet too).

I'd hope we'd do some user testing around how people use this and explore how used it is on tablet before disabling it.

Maybe for the purpose of this patch expanding by default might make more sense and be less controversial?

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0