Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

This patch appears to really be doing two contradicting things:
1. Making the default value of $wgMinervaBottomLanguageButton false
2. Removing support for $wgMinervaBottomLanguageButton

If the goal is (1), then the tests should not be removed, but the config variable should be forced on during relevant tests. If the goal is (2), then the config variable and all references therein should be removed. In addition, any tests for the "new" language switcher icon/overlay should be merged into the existing language test files.

To clarify, we shouldn't remove tests for features we still technically support just because we expect them to be disabled.

Patch-set: 1