Update patch set 1
Patch Set 1: Code-Review+1

To add a little history, this test step dates back to when we used to re-render the page via ajax after an edit. Since we don't do that now - we reload the page - this test step is not very useful and imo not adding any value so I am okay with removing it. In that sense it is unrelated to testing the MinervaBottomLanguageButton feature.

MinervaBottomLanguageButton will be removed next week (see other open patchset - https://gerrit.wikimedia.org/r/315318) so let's not create unnecessary work for ourselves.

@jhobs with that in mind are you happy for me to +2?

Patch-set: 1