Update patch set 1
Patch Set 1: Code-Review+1

Yes, thank you for the context. I think a FIXME could probably be added to make it explicit that we are planning to remove this feature next week in case anyone comes upon it before then, but it's not a necessity.

Patch-set: 1
Reviewer: Jhobs <1948@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1