Update patch set 2
Patch Set 2:

No rush with getting this merged. I didn't want to interfere with the sprint goal and I felt it was better to start this clean up after the language button was pushed which seems more important.

Hopefully the dependency will get merged next week, if we decide not to, I'll rebase against master when that happens.

Feel free to change to a +2 - it will then merge automatically when the dependency is merged.

Patch-set: 2
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0