Update patch set 3
Patch Set 3:

Whether it's important or not is up for debate :)

I was interested in what reactions were if, any, but really the more interesting patch that I was hoping for some feedback on was https://gerrit.wikimedia.org/r/#/c/315426/2

I'm having a conversation with Joaquin around his mfui library and I've been a bit frustrated with how difficult it is to make an overlay so I was exploring a different approach.

https://github.com/joakin/mfui/issues/7 has more background, but essentially we need to make a decision about how we want to use templates going forward and if we want to allow overriding of them.

No urgency to review/merge this at current time.

Patch-set: 3
Label: Verified=0