Unify the logic to brand the search functionality across mobile modes
In order to decide on the branding message, take MobileContext mode
rather than the Skin mode into account. In the long run this will
help us remove the beta skin/template. We will be able to enable a
set of features and call it a beta mode without assigning a mode
to any skin.

Bug: T148197
Change-Id: I1e30cc83c126f2b63b65504e894dd473b655fc46