Update patch set 1
Patch Set 1:

Do we have any idea what the overhead of calling MobileContext:: singleton is?

Would it make sense to cache the result?

Also it's worth noting that with this change we would call it from a desktop context when viewing desktop version of minerva.

Would it make sense to use setters and getters in the hooks or to pass this information in the constructor of the skin?

Patch-set: 1
Label: Verified=0