Update patch set 4
Patch Set 4:

I see now that the same class is re-used for something completely different by the icons module to inverse the arrow. I guess that means this does indeed work. But it seems odd to apply open-block to the heading since any styles for that class directly would be intended for the section, not the heading. It works right now because styles for open-block target ". collapsible-block.open-block" only.

Patch-set: 4
Label: Verified=0