Update patch set 9
Patch Set 9:

The benefit of adding JS to skin over to MobileFormatter is that MobileFormatter is used to return API results too. That's why $scriptsEnabled was introduced. This complexity can be avoided. Also, the response from the API (that the apps use) will be the same as the response we get on the mobile site - caching benefit. We also won't have to mix JS with HTML - separation of concerns. Possibly other benefits too, but those are good enough reasons.

Patch-set: 9