Update patch set 9
Patch Set 9:

In an ideal world the entire skin would be api driven. I agree with baha that this would be a better approach but when considering implementation this was a non-starter.


The onscript attributes have to be added in MobileFormatter. I therefore decided that the script tag was better added here too, especially given other skins can use MobileFormatter.

In an ideal world MobileFormatter would provide a structured json response like the mobile content service does. I'm planning a conversation around this after the all hands and when this is the case none of this markup will be needed inside MobileFormatter

Patch-set: 9