Update patch set 9
Patch Set 9:

With respect to add the event handler on the body tag to avoid needing to use the onclick attributes and using an event delegated approach (sorry only just thoroughly groked that):

I'm not too comfortable with this approach so late in the sprint as I must confess I didn't consider this method during prototyping. Here are some concerns off the top of my head:

* There are many things that could preventDefault or stopPropagation and interfere with this approach, Central Notice comes to mind.
* IE 6-8  and possibly other browsers do not have an `event.target` property. This was why I made use of ids to avoid having to deal with this (remember if this code fails the page becomes inaccessible))
* We'd also have to unbind that event inside the toggle code to avoid weird side effecs. How would we do that cleanly?

I'm happy to continue this conversation on Phabricator, but in an ideal world I'd like MobileFormatter to behave like the MCS and allow the skin to completely take control of section collapsing and the HTML of section headings. Toggling is a Minerva only thing.

Patch-set: 9