Update patch set 1
Patch Set 1:

Thanks for this! I've seen this but I'd like to read through all the changes to code to understand what rules are being disabled and why (so we can track upstream bugs) and what the code changes relate to (whether it's just more strict whitespacing or new rules we may want to okay).

If someone wants to do that and update the commit message that would be helpful!

Please can I ask that we do not merge until that's done.

Patch-set: 1