Avoid a flash of unstyled content on most tablet devices
To avoid a flash of unstyled content for tablet devices we need
to use a css selector that is true at page load but not after the
toggling code has loaded. Rather than adding unnecessary additional
artifacts I've decided to make use of the onclick attribute to identify
this state given it is removed by the toggling code.

Note that any browsers that do not support media queries or css
sibling selectors will still have the flash of unstyled content.

It seems that browsers in Grade A support are generally supported by
this combination of selectors.

Bug: T150839
Change-Id: I60b29a09feba626ed70538e7ba88c6ab96d463aa