Scope talk styles to talk overlay
Currently the talk styles can influence the
styles of other overlays, but they are not
in mobile.overlays

This scopes the rules to the talk overlay to avoid
inconsistent renders of overlays depending on whether
the talk overlay has been opened before.

The styles are no longer loaded on page load.

Changes:
* The talk-overlay class is no longer added to overlay-content
element. This was inconsistent from usages in other overlays.
* The mobile-frontend-editor-licensing messages are no longer
loaded twice given they are loaded in mobile.startup.
* The messages for mobile-frontend-talk-overlay-header and
mobile-frontend-talk-add-overlay-submit which are loaded in
mobile.talk.overlays are no longer added twice.

Change-Id: Ia9bedce5d77461ae5c99673637051bdcd4f125ad