Fix for ios talk overlay when virtual keyboard open
The emulated scroll code for older ios browsers
ppears to be breaking newer versions of iOS.

The link to the apple discussion is not helpful. There are no replies
on how to fix it and it requires a login to access but it seems simply
by not running the 3 lines of code the right behaviour is present in
the more recent versions of ios I tested.

In ios9 it was interfering with the scroll top set natively masking
the textarea from view.

I've pulled this code out from the postRender method to make it clearer
what it is responsible for. This wasn't 100% clear beforehand.

Bug: T149363
Change-Id: I6856303b0982bec843b3d8f864b1f111698f6439