Update patch set 1
Patch Set 1:

(2 comments)

This can not happen in production. It's not possible for a user to accidentally trigger this code without a title. There is always a title when you are on a special page, when you edit or preview a page, … always.

But this can happen in tests when a context object was created as a dummy or mock without a title. Unrelated code that happens to trigger this hook should just not crash then.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0