Update patch set 1
Patch Set 1: -Code-Review

(4 comments)

Removing -1 because your solution works. I was worried that we're making the already complex toggling code more complex. It's weird how we add 'collapsible-heading' with JS, and 'collapsible-section' in PHP. Maybe it's time to revisit our toggling code and make it robust rather than fix issues with it once we become aware of them.

Patch-set: 1