Update patch set 1
Patch Set 1:

One more finding, it doesn't happen when I browse in iPad mode from beginning. After changing the device to iPad and refreshing the page it works properly. 
But it doesn't work in different-tablet like resolutions. For example when emulating Nexus10 or just when viewport is wider than 1000px

Patch-set: 1