Update patch set 2
Patch Set 2:

(1 comment)

It's not gonna be done on every page view? It's adding a link, not the contents of the API call...

The API is well cached at the varnish layer, if appropriately configured, and headers sent by the API

You can override the printer in the API so it always returns JSON, as many other extension and similar API modules do (such as RSD which is an XML based format). It's fine to do, just document it.

Patch-set: 2