Update patch set 3
Patch Set 3:

A static JSON file wouldn't have the correct name value which IMHO is the most important information (the colors are only cosmetic).

Also note that the Chrome install banner only appears when the webapp manifest sets the display mode to standalone (+ a bunch of other conditions) which is not the case here.

Patch-set: 3
Label: Verified=0