Update patch set 7
Patch Set 7:

My concern with shipping additional ico8 additional icons is we will double the size of the stylesheet for the purpose of an icon that only shows when a link is active or hovered (which only happens when the icon is clicked on a mobile device). Given the icon only changes on hover I think this is a sensible reasonable approach. The browser support is good for the majority of devices that will use the mobile site and for those that do not they will simply always see a black icon and have no reason to believe their experience is broken. I agree the matrix and inlined SVG may be hard to debug but Id like to give this a go and revisit later if necessary. This is a small contained use case.

Patch-set: 7
Label: Verified=0