Update patch set 8
Patch Set 8:

Valid points. I'd love to see these captured in a task.

That said, given the problem at hand, we should attempt the fasts sensible solution. We should also evaluate how well the matrix solution scales. That said, the hovering effect is mostly visible when the user taps on an item on mobile devices. If the Internet is fast enough, that won't be noticeable too much.

Patch-set: 8