Hygiene: breakup mega-function in MobileView API
In order to make this block of code testable and readable we
should break it out into a separate method.

Changes:
* I've renamed parseSections to the more meaningful
getRequestedSectionIds as the name parseSections is extremely
misleading.
* I've broken out parseSectionsData and documented it as its
behaviour is not what I expected - the resulting array
has a text property rather than adding the text property directly to
the section it relates to. I'd like to revisit this behaviour in future.

Change-Id: I8045d5eb6cc5d1ddce7bc344d99566aa648f70fd