Update patch set 2
Patch Set 2: Code-Review-1

It was my understanding in standup you said we were going to remove this flag altogether as there doesn't seem to be a need for a feature flag at all now (and if there is that should live in CentralNotice). Please update ticket if that's not the case.

Patch-set: 2
Label: Code-Review=-1